+++ resources/views/address/create.blade.php
@@ -7,7 +7,7 @@
         <x-laraknife.panels.create title="{{ __('Creation of an Address') }}">
             <x-laraknife.forms.combobox position="first" name="addresstype_scope" label="Addresstype" :options="$optionsAddresstype"
                 width2="4" />
-            <x-laraknife.forms.combobox position="last" name="owner_id" label="Owner" :options="$optionsOwner" width2="4" />
+            <x-laraknife.forms.combobox position="last" name="person_id" label="Person" :options="$optionsPerson" width2="4" />
             <x-laraknife.forms.string position="first" name="name" label="Email/Phone"
                 value="{{ $context->valueOf('name') }}" width2="4" />
             <x-laraknife.forms.string position="last" name="priority" label="Priority"