+++ resources/views/address/show.blade.php
@@ -8,7 +8,7 @@
         @endif
         <x-laraknife.panels.show title="{{ __($mode !== 'delete' ? 'A Address' : 'Deletion of a Address') }}" mode="{{$mode}}">
             <x-laraknife.forms.combobox position="first" name="addresstype_scope" label="Addresstype" :options="$optionsAddresstype" width2="4" attribute="readonly"/>
-            <x-laraknife.forms.combobox position="last" name="owner_id" label="Owner" :options="$optionsOwner" width2="4" attribute="readonly"/>
+            <x-laraknife.forms.combobox position="last" name="person_id" label="Person" :options="$optionsPerson" width2="4" attribute="readonly"/>
             <x-laraknife.forms.string position="first" name="name" label="Name" value="{{ $context->valueOf('name') }}" width2="4" attribute="readonly" />
             <x-laraknife.forms.string position="last" name="priority" label="Priority" value="{{ $context->valueOf('priority') }}" width2="4" attribute="readonly" />
             <x-laraknife.forms.text position="alone" name="info" label="Info" value="{{ $context->valueOf('info') }}" width2="10" rows="2" attribute="readonly" rows="2" />