+++ resources/views/location/index.blade.php
@@ -19,12 +19,12 @@
                 <thead>
                     <tr>
                         <th></th>
+                        <th sortId="person">{{ __('Person') }}</th>
                         <th sortId="country">{{ __('Country') }}</th>
                         <th sortId="zip">{{ __('Zip') }}</th>
                         <th sortId="city">{{ __('City') }}</th>
                         <th sortId="street">{{ __('Street') }}</th>
                         <th sortId="additional">{{ __('Additional') }}</th>
-                        <th sortId="owner">{{ __('Owner') }}</th>
                         <th></th>
                     </tr>
                 </thead>
@@ -32,12 +32,12 @@
                     @foreach ($records as $location)
                         <tr>
                             <td><x-laraknife.icons.change-record module="location" no="{{ $location->id }}" /></td>
+                            <td>{{ $location->person }}</td>
                             <td>{{ $location->country }}</td>
                             <td>{{ $location->zip }}</td>
                             <td>{{ $location->city }}</td>
                             <td>{{ $location->street }}</td>
                             <td>{{ $location->additional }}</td>
-                            <td>{{ $location->owner }}</td>
                             <td><x-laraknife.icons.delete-record module="location" no="{{ $location->id }}" /></td>
                         </tr>
                     @endforeach