+++ templates/database/migrations/2024_11_26_075255_create_addresses_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('info')->nullable();
             $table->integer('priority');
-            $table->foreignId('owner_id')->nullable()->references('id')->on('users');
+            $table->foreignId('person_id')->nullable()->references('id')->on('persons');
         });
     }